Add instructions.json for Popit countries

+++ rakefile_popit.rb
@@ -4,22 +4,29 @@ require_relative 'rakefile_common.rb'
 desc "Reload the source data"
 task :raw => 'popit.json'
 
+
+@SOURCE_DIR        = 'sources/popit'
+@POPIT_RAW_FILE    = @SOURCE_DIR + '/raw.json'
+@INSTRUCTIONS_FILE = @SOURCE_DIR + '/instructions.json'
+
+CLOBBER.include(@POPIT_RAW_FILE)
+
+if File.exist? @INSTRUCTIONS_FILE
+  @instructions = json_load(@INSTRUCTIONS_FILE)
+  raise "No `source` in instructions.json" unless @instructions.key? :source
+end
+
 def popit_source 
-  @POPIT_URL || "https://#{@POPIT || @DEST}.popit.mysociety.org/api/v0.1/export.json"
+  "https://%s.popit.mysociety.org/api/v0.1/export.json" % @instructions[:source]
 end
 
 namespace :raw do
-
-  @POPIT_RAW_FILE = 'sources/popit/raw.json'
-
   file @POPIT_RAW_FILE do
     File.write(@POPIT_RAW_FILE, open(popit_source).read) 
   end
-
 end
 
 namespace :whittle do
-
   @POPIT_CLEAN_FILE = 'clean.json'
 
   task :load => @POPIT_RAW_FILE do
@@ -32,6 +39,5 @@ namespace :whittle do
       end
     }, { symbolize_names: true })
   end
-
 end
